Update js/app.js
Changed a few minor things, still working on a few more code changes that I would like to have in there. Need to add functionality for drop down menus and populating them from Cardio Pulse.

I do have a small hichup, I am not sure how we are going to identify the web server here, The way our backend database works is that each instance of the application or each site has its own application and will have its own web server that will have a unique address to that site. So I will need some way to give the application the correct address to make the web request to. If you have any ideas please let me know.

When we make this application so that it can update from the website with out having to rebuild? I realize that is basically just making a browser but that way we could update it with out having customers have to rebuild, not that we plan on releasing too many versions of this after the changes.